+++ terrain/voxel_mesh_block.h
@@ -67,6 +67,8 @@ public:
 	// Collisions
 
 	void set_collision_mesh(Vector<Array> surface_arrays, bool debug_collision, Spatial *node);
+	void set_collision_layer(int layer);
+	void set_collision_mask(int mask);
 	void drop_collision();
 	// TODO Collision layer and mask